Move collaborators under the `client.domains` namespace.
