+++ src/main/java/com/dnsimple/Domains.java
@@ -5,6 +5,10 @@ import com.dnsimple.response.GetDomainResponse;
 import com.dnsimple.response.CreateDomainResponse;
 import com.dnsimple.response.DeleteDomainResponse;
 import com.dnsimple.response.ResetDomainTokenResponse;
+import com.dnsimple.response.ListCollaboratorsResponse;
+import com.dnsimple.response.AddCollaboratorResponse;
+import com.dnsimple.response.RemoveCollaboratorResponse;
+import com.dnsimple.exception.DnsimpleException;
 import com.dnsimple.response.EnableDnssecResponse;
 import com.dnsimple.response.DisableDnssecResponse;
 import com.dnsimple.response.GetDnssecResponse;
@@ -112,6 +116,63 @@ public interface Domains {
    * @throws IOException Any IO errors
    */
   public ResetDomainTokenResponse resetDomainToken(String accountId, String domainId) throws DnsimpleException, IOException;
+
+  // Collaborators
+
+  /**
+   * Lists the collaborators in the domain.
+   *
+   * @see <a href="https://developer.dnsimple.com/v2/domains/collaborators/#list">https://developer.dnsimple.com/v2/domains/collaborators/#list</a>
+   *
+   * @param accountId The account ID
+   * @param domainId The domain ID
+   * @return The list collaborators response
+   * @throws DnsimpleException Any API errors
+   * @throws IOException Any IO errors
+   */
+  public ListCollaboratorsResponse listCollaborators(String accountId, String domainId) throws DnsimpleException, IOException;
+
+  /**
+   * Lists the collaborators in the account.
+   *
+   * @see <a href="https://developer.dnsimple.com/v2/domains/collaborators/#list">https://developer.dnsimple.com/v2/domains/collaborators/#list</a>
+   *
+   * @param accountId The account ID
+   * @param domainId The domain ID
+   * @param options A Map of options to pass to the collaborators API
+   * @return The list collaborators response
+   * @throws DnsimpleException Any API errors
+   * @throws IOException Any IO errors
+   */
+  public ListCollaboratorsResponse listCollaborators(String accountId, String domainId, Map<String,Object> options) throws DnsimpleException, IOException;
+
+  /**
+   * Add a collaborator to a domain.
+   *
+   * @see <a href="https://developer.dnsimple.com/v2/domains/collaborators/#add">https://developer.dnsimple.com/v2/domains/collaborators/#create</a>
+   *
+   * @param accountId The account ID
+   * @param domainId The domain ID
+   * @param attributes A Map of attributes for constructing the domain
+   * @return The add collaborator response
+   * @throws DnsimpleException Any API errors
+   * @throws IOException Any IO errors
+   */
+  public AddCollaboratorResponse addCollaborator(String accountId, String domainId, Map<String,Object> attributes) throws DnsimpleException, IOException;
+
+  /**
+   * Remove a collaborator from a domain.
+   *
+   * @see <a href="https://developer.dnsimple.com/v2/domains/collaborators/#remove">https://developer.dnsimple.com/v2/domains/collaborators/#remove</a>
+   *
+   * @param accountId The account ID
+   * @param domainId The domain ID
+   * @param collaboratorId The collaborator ID
+   * @return The remove collaborator response
+   * @throws DnsimpleException Any API errors
+   * @throws IOException Any IO errors
+   */
+  public RemoveCollaboratorResponse removeCollaborator(String accountId, String domainId, String collaboratorId) throws DnsimpleException, IOException;
 
   // DNSSEC
