+++ src/main/java/com/dnsimple/endpoints/EndpointAdapter.java
@@ -2,7 +2,6 @@ package com.dnsimple.endpoints;
 
 import com.dnsimple.Accounts;
 import com.dnsimple.Certificates;
-import com.dnsimple.Collaborators;
 import com.dnsimple.Contacts;
 import com.dnsimple.Domains;
 import com.dnsimple.Identity;
@@ -20,8 +19,6 @@ public interface EndpointAdapter {
   public Accounts accounts();
 
   public Certificates certificates();
-
-  public Collaborators collaborators();
 
   public Contacts contacts();
